+++ app/_actions/config-assistant-for-project.ts
@@ -49,7 +49,7 @@ export async function configAssistantForProject(
   project: Project,
   fileInfos: ProjectDataReference[],
   user: User,
-  org: Organization | OrgPartDeux,
+  org: Organization | OrgPartDeux
 ): Promise<Assistant> {
   const session = await auth()
 
@@ -87,6 +87,7 @@ export async function configAssistantForProject(
       const updatedAssistant = await updateAssistantPromptAndFiles(
         fileInfos,
         existingAssistant,
+        undefined
       )
       await setCachedProjectUserFileInfos(project.name, user, fileInfos)