+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added advanced project configuration button which includes old checkbox selection menu and project description
 - Changed wording on status texts in UI to clearer represent purposes
 - Cleaned up UI
+- whenever user navigates to the main page, it will automatically set the user's org to be the first of the fetched orgs. otherwise it will stay on the create orgs screen
 
 ### Bug Fixes